Final file 99%

+++ server/api/announcement/createAnnouncement.post.ts
@@ -11,8 +11,10 @@ export default defineEventHandler(async (event) => {
 	const type = formData.find((field: any) => field.name === "kontenType")
 	const cat_id = formData.find((field: any) => field.name === "category")
 	const file = formData.find((field: any) => field.name === "file")
+	const fileUrl = formData.find((field: any) => field.name === "previewUrl")
+	const saveType = formData.find((field: any) => field.name === "type")
 
-	if (!title || !type || !cat_id || !file) {
+	if (!title || !type || !cat_id || !file || !saveType) {
 		return {
 			statusCode: 400,
 			body: {
@@ -26,6 +28,7 @@ export default defineEventHandler(async (event) => {
 		type: type.data.toString(),
 		cat_id: JSON.parse(cat_id.data).id,
 		cat_name: JSON.parse(cat_id.data).name.replace(" ", ""),
+		fileUrl: fileUrl ? fileUrl.data.toString() : "",
 		file: file,
 	}
 
@@ -42,43 +45,62 @@ export default defineEventHandler(async (event) => {
 
 			const newNameField = announcement.an_id
 			const kategori = dataAnnouncement.cat_name
-			const ext = path.extname(file.filename).toLowerCase()
+			const extCek = path.extname(announcement.an_id).toLowerCase()
+			let ext = ""
+			if (extCek === "") {
+				ext = ".png"
+			} else {
+				ext = extCek
+			}
+
 			const tempDir = tmpdir()
-			const tempFilePath = path.join(tempDir, file.filename)
+			const tempFilePath = path.join(tempDir, announcement.an_id)
 			const resourcesDir = path.resolve(`resources/${kategori}`)
-            const accFilePath = path.join("/_nuxt","resources", `${kategori}`, `${newNameField}${ext}`)
+			const accFilePath = path.join("/_nuxt", "resources", `${kategori}`, `${newNameField}${ext}`)
 			const newFilePath = path.join(resourcesDir, `${newNameField}${ext}`)
+			if (saveType.data === "file") {
+				try {
+					await fs.writeFile(tempFilePath, file.data)
 
-			try {
-				await fs.writeFile(tempFilePath, file.data)
+					await fs.mkdir(resourcesDir, { recursive: true })
 
-				await fs.mkdir(resourcesDir, { recursive: true })
-
-				if ([".jpg", ".jpeg", ".png", ".webp"].includes(ext)) {
-					// Compress and save the image using sharp
-					await sharp(tempFilePath)
-						.resize(800) // Resize to a max width of 800px
-						.toFormat("webp") // Convert to webp for better compression
-						.toFile(newFilePath)
-				} else if ([".mp4", ".mov", ".avi"].includes(ext)) {
-					await new Promise((resolve, reject) => {
-						hbjs.spawn({
-							input: tempFilePath, // Use the temporary file path
-							output: newFilePath,
-							preset: "Very Fast 720p30", // Adjust the preset as needed
+					if ([".jpg", ".jpeg", ".png", ".webp"].includes(ext)) {
+						// Compress and save the image using sharp
+						await sharp(tempFilePath)
+							.resize(800) // Resize to a max width of 800px
+							.toFormat("webp") // Convert to webp for better compression
+							.toFile(newFilePath)
+					} else if ([".mp4", ".mov", ".avi"].includes(ext)) {
+						await new Promise((resolve, reject) => {
+							hbjs.spawn({
+								input: tempFilePath, // Use the temporary file path
+								output: newFilePath,
+								preset: "Very Fast 720p30", // Adjust the preset as needed
+							})
+								.on("error", reject)
+								.on("end", resolve.bind(null, null))
 						})
-							.on("error", reject)
-							.on("end", resolve.bind(null, null))
-					})
-				} else {
-					throw new Error("Unsupported file type")
-				}
+					} else {
+						throw new Error("Unsupported file type")
+					}
 
-				// Optionally, delete the temporary file
-				await fs.unlink(tempFilePath)
-			} catch (error) {
-				console.error("Error processing the file:", error)
-				throw new Error("Failed to compress and save the file")
+					// Optionally, delete the temporary file
+					await fs.unlink(tempFilePath)
+				} catch (error) {
+					console.error("Error processing the file:", error)
+					throw new Error("Failed to compress and save the file")
+				}
+			} else {
+				try {
+					const base64Data = dataAnnouncement.fileUrl.replace(/^data:image\/png;base64,/, "")
+					const binaryData = Buffer.from(base64Data, "base64")
+					await fs.writeFile(newFilePath, binaryData)
+
+					await fs.mkdir(resourcesDir, { recursive: true })
+				} catch (error) {
+					console.error("Error processing the file:", error)
+					throw new Error("Failed to compress and save the file")
+				}
 			}
 
 			await prisma.announcement.update({
